Extract `PdfNavigatorDocument` and PDFium
