+++ readium/adapters/pdfium/pdfium-document/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Readium Foundation. All rights reserved.
+ * Copyright 2022 Readium Foundation. All rights reserved.
  * Use of this source code is governed by the BSD-style license
  * available in the top-level LICENSE file of the project.
  */
@@ -13,7 +13,10 @@ plugins {
 }
 
 android {
+    resourcePrefix = "readium_"
+
     compileSdk = 32
+
     defaultConfig {
         minSdk = 21
         targetSdk = 32
@@ -23,18 +26,18 @@ android {
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
     }
-    testOptions {
-        unitTests.isIncludeAndroidResources = true
-    }
     kotlinOptions {
+        jvmTarget = "1.8"
         freeCompilerArgs = freeCompilerArgs + "-opt-in=kotlin.RequiresOptIn"
-        allWarningsAsErrors = true
     }
     buildTypes {
         getByName("release") {
             isMinifyEnabled = false
             proguardFiles(getDefaultProguardFile("proguard-android.txt"))
         }
+    }
+    buildFeatures {
+        viewBinding = true
     }
 }
 
@@ -44,7 +47,7 @@ afterEvaluate {
             create<MavenPublication>("release") {
                 from(components.getByName("release"))
                 groupId = "com.github.readium"
-                artifactId = "readium-streamer"
+                artifactId = "readium-adapter-pdfium-document"
                 artifact(tasks.findByName("sourcesJar"))
                 artifact(tasks.findByName("javadocsJar"))
             }
@@ -57,30 +60,14 @@ dependencies {
 
     api(project(":readium:shared"))
 
-    implementation("androidx.appcompat:appcompat:1.4.1")
-    @Suppress("GradleDependency")
+    implementation("androidx.core:core-ktx:1.8.0")
+    implementation("com.github.barteksc:pdfium-android:1.8.2")
     implementation("com.jakewharton.timber:timber:5.0.1")
-    implementation("com.github.edrlab.nanohttpd:nanohttpd:master-SNAPSHOT") {
-        exclude(group = "org.parboiled")
-    }
-    implementation("com.github.edrlab.nanohttpd:nanohttpd-nanolets:master-SNAPSHOT") {
-        exclude(group = "org.parboiled")
-    }
-    //AM NOTE: conflicting support libraries, excluding these
-    implementation("com.mcxiaoke.koi:core:0.5.5") {
-        exclude(module = "support-v4")
-    }
-    // useful extensions (only ~100k)
-    implementation("com.mcxiaoke.koi:async:0.5.5") {
-        exclude(module = "support-v4")
-    }
-    implementation("joda-time:joda-time:2.10.14")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.1")
 
-    // Tests
     testImplementation("junit:junit:4.13.2")
-    testImplementation("org.assertj:assertj-core:3.22.0")
-    testImplementation("org.robolectric:robolectric:4.7.3")
+
     androidTestImplementation("androidx.test.ext:junit:1.1.3")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.4.0")
 }
