+++ readium/adapters/pdfium/pdfium-document/src/main/AndroidManifest.xml
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest package="org.readium.adapters.pdfium.document" />
+
