+++ readium/adapters/pdfium/pdfium-navigator/src/main/res/layout/readium_pspdfkit_fragment.xml
