--- readium/navigator/src/main/res/layout/r2_pdf_activity.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/r2_pdf_navigator"
-        android:name="org.readium.r2.navigator.pdf.PdfNavigatorFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
