+++ readium/navigator/src/main/res/values/ids.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <item name="readium_pdf_container" type="id" />
+</resources>
